requirements.txt

blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dash==3.0.4
dash-bootstrap-components==2.0.3
Flask==3.0.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==1.42.0
nest-asyncio==1.6.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pip==25.1.1
plotly==6.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
retrying==1.3.4
setuptools==80.9.0
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.23.0
psycopg2-binary==2.9.10
dotenv==0.9.9